Slider, Audio, and Asset Tweaks
Removed MetaAssets from gitignore so that all assets are backed up. Sliders have been very inconsistent based on game window resolution but it should be largely fixed. In addition I added a sqrt operation to the volume sliders, meaning that resulting volume levels are much closer to expected based on where the sliders are positioned